[manuf] update individualization to exclude AST data from OTP digest
We purposely exclude the AST configuration data field of the CreatorSwCfg
partition from the digest calculation because this could be different
per chip and we do not want to it to be part of the foundation for the
UDS keys if the OWNER_SW_CFG_ROM_KEYMGR_OTP_MEAS_EN switch is set to
enabled. If this field is intended to be set for a given SKU, during
personalization we need to be able to inject the expected OTP SW
partition measurement into the manifest of the perso image so that the
same CreatorRootKey keymgr attestation binding value computed in the
field by the ROM is the same as the one used during perso when all the
CreatorSwCfg fields have not yet been set.

Signed-off-by: Tim Trippel <ttrippel@google.com>
(cherry picked from commit 832c18940a45b70155c48b429e9c288dfb9653d4)