[rom_ext] Order the ePMP as expected by customers
Traditionally, we placed the RvDm region in ePMP entry 13 and the
memory-mapped flash in region 12.

This was reversed because it would allow for one more ToR region when RvDM
is not in use, however downstream customers applications may depend on a
specific ePMP configuration when their firmware starts.

The customer is free to re-arrange ePMP as they wish, so a configuration
with an extra ToR entry is still possible as long as the customer programs
the proper ePMP arragement.

Signed-off-by: Chris Frantz <cfrantz@google.com>
(cherry picked from commit a9a57fae05eba95c3db3b2b3ccc9871ead45e30d)