[alert,fpv] Rephrase the handshake assertion in async rxtx alert fpv
The old version of this assertion had a bit of a hole in it. Digging
into the SVA, we see:

    $rose(prim_alert_rxtx_async_tb.alert_pd)   ##[3:6]
    $rose(prim_alert_rxtx_async_tb.ack_pd)     &&
    $stable(prim_alert_rxtx_async_tb.alert_pd) ##[3:6] ....

This says "alert_p rises and then in between 3 and 6 cycles ack_p
rises and alert_p is stable at that point". That's not quite what we
want!

We really want to say "alert_p rises and then stays stable for up to 6
cycles at which point ack_p rises."

I've rephrased it slightly and also been a bit more careful about
differential signals and who is seeing what. Now, the FullHandshake_S
sequence is explicitly from the alert sender's point of view.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>