[alert,fpv] Rather simplify some of the async alert code assertions
This has an equivalent meaning, except that I added an extra disable
condition to AlertReqAck_A. That extra condition shouldn't really be a
problem: this is saying that if I ask an alert sender to send an alert
then it will eventually get a response. Of course, that's not true if
someone cut the wires! But we don't care because the receiver will
eventually see a failed ping.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>