[topgen] Support external TL-UL connections
Introduced some attributes that allow for an external TL-UL port to be added to a particular top.

For a `host` port, it requires defining the external device as a module in the top configuration
file; include the device in the address space for the top and include the following additional
fields:
- `external`, indicating that it is a external device
- `external_size`, indicating the byte size of the address space of the
  device

For a `device` port, adding the node in the corresponding `xbar` with
the `external` field will generate the corresponding port connection.

Signed-off-by: Luis Martinez <luism@zerorisc.com>