[c++,sw] Add clang support
This change makes the necessary modifications to unit tests
and build files to enable a full build of the project with Clang.

This, in turn, enables the whole project to be fuzzed via Google
FuzzTest, which requires the use of Clang.

These changes are very similar to those on earlgrey_1.0.0 as part
of https://github.com/lowRISC/opentitan/pull/27595. However, this
is not a cherry-pick due to differences in test and BUILD files.

Signed-off-by: Collin MacDonald <cmacd@google.com>