[dv] Reorder dv_base_env_cfg and add documentation comments
The function that was called create_ral_models can be dramatically
simplified by defining a create_ral_model function to do the body and
looping with that. Unfortunately, the name is a bit close to
create_ral_by_name, so I've renamed the two local functions to
make_ral_model(s).

Also note that some of the documented items currently only apply to
CIP-related objects. For example, en_scb_mem_chk is only used for TL
transactions. But the general config option might make sense more
generally, so it probably still belongs in this config object.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>