[otbn] Added error flags for kmac req caused by incorrect message sizes and preventing lockups
Added status flags for oversized and undersized messages.
Added logic to prevent stalls and artificially complete transactions.
For undersized transactions a last signal is injected to complete the
message and receive a digest from KMAC, with an appropriate error status.
For oversized messages the remainder is flushed from the MSG FIFO and
the appropriate error status is reported.

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>