[sw,cov] Disable segment allocation for coverage bss sections
The `__llvm_prf_cnts` section is stripped during conversion to `.bin` files
to save firmware size, as it's always filled with `0xff`.

However, unlike standard BSS, this section contains non-zero initial contents,
leading the compiler to mark it as `PROGBITS` and loadable. This PR modifies
the linker script to explicitly allocate this section to `NONE` segment,
preventing the `qemu flashgen` tool from loading it.

The special `NONE` segment tells the linker to not put the section in any
segment at all.

Change-Id: I917a3e97e0afcd17c3231b0cfb5ce61c870ce292
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>