[ci] Fix bug in bitstream checkout for FPGA tests
The bitstream checkout process for FPGA tests uses the uploaded bitstream
artifact in GHA. However, when bitstreams_workspace.py is run prior to the FPGA
tests (without network fetch), it expects that archive to be in the format used
for the GCP bitstream cache. The former lacks global manifest.json file with the
commit hash (separate from the per-bitstream manifest.json).

This commit fixes this defect by preceding the FPGA tests with a build of a
GCP-capable cache archive that is uncompressed into the cache directory.

Signed-off-by: Ryan Torok <rtorok@zerorisc.com>