[hw,ac_range_check,rtl] Implement log_status as hwext to allow clearing it
Previously, log_status is implemted as an ordinary register. log_status.log_clear
was used to detect rising-edge writes to it, which should clear the log_status.
However, clear_log is always tied to 0 (.de=1, .d=0), causing that it is always
cleared and rising edges cannot be detected.

To fix that, the register is implemented as hwext (as in other IPs where the rising
edge matters). Other fields need to be implemented as ordinary register though manually.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>