[fisim] Add GCM test and cleanup
Add the GCM instruction skip test.
Since GCM is purely in SW, fisim had to be adapted to tackle it as GDB
crashes with it. The tracing is adapted such that functions can be
skipped. The exception handling is also expanded to deal with timeout
errors from gdb when reinitializing.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>