[rust] replace `Rc::clone` with `.clone()`
`Rc::clone` was previously a recommended way of cloning a reference-counted
pointer however this is not the case anymore (relevant clippy warning has been
moved to pedantic instead of enabled by default quite a while back, for example).

We use mostly `Rc<dyn Trait>` which can never be deeply copied anyway, as `Clone`
is not dyn-safe.

This change makes it easier to replace `Rc` with `Arc` in the future.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>