[opentitanlib] use a constructor to create `UartConsole`
Currently all fields of `UartConsole` is exposed as public and users can
use struct syntax to create this. This is bad for knowing what's actually
public and what's not. Add a `new` method and use it in tests.

Fields that are not modified outside `UartConsole` and thus be made private.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>