[fisim] Add ROM_EXT GDB testing
Add a new test target for instruction skip testing of the ROM_EXT.
The ROM_EXT testing is done with RMA OTP but setting and locking SECRET2
and setting the RMA spin cycles to maximum.
A wrongly signed BL0 image is created and we create separate targets for
manifest images to allow testing of various malicious flash inputs.

The GDB testing halts the CPU in the RMA spin cycles, jumps the loop,
then traces the rom_ext_try_next_stage function and inserts instruction
skips with a corrupted BL0 image.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>