[FiSim] Add unit tests for FiSim
The FiSim framework typically takes a couple of hours for a regular
instruction skip test. In order to have a version of this framework in
CI, we create small unit tests showing secure coding styles which are
then testing via instruction skips.

Info on those guidelines has been added in the firmware_gdb.c file and
we refer to this file in the README.

In case this test catches in CI:
This is a small binary which is tested via OpenOCD/GDB specifically for
an RMA ROM on the CW340 FPGA.
Please adapt the code/raise an issue/contact someone, since the other
tests in //sw/host/penetrationtests/python/fi/gdb_testing would need to be
adapted as well. However, these tests take a long time, hence they are
out of CI.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>