[pentest] Add clober list to ASM based trigger
For the pentests, we use an ASM based trigger to minimize to code overhead
during the trigger window.

However, as this ASM snippets modify register contents, we should add the
registers we are modifying to the clobber list. With this, we tell the
compiler which of the registers we are modifying during the ASM trigger
snippet.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>